+++ src/main/java/com/bawnorton/trimica/client/mixin/render/ShieldSpecialRendererMixin.java
@@ -1,21 +1,12 @@
 package com.bawnorton.trimica.client.mixin.render;
 
-import com.bawnorton.trimica.client.TrimicaClient;
-import com.bawnorton.trimica.client.palette.TrimPalette;
-import com.bawnorton.trimica.client.texture.DynamicTrimTextureAtlasSprite;
-import com.bawnorton.trimica.compat.Compat;
+import com.bawnorton.trimica.api.client.TrimicaClientApi;
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.blaze3d.vertex.VertexConsumer;
 import dev.kikugie.fletching_table.annotation.MixinEnvironment;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.ShieldModel;
-import net.minecraft.client.renderer.LightTexture;
 import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.renderer.special.ShieldSpecialRenderer;
 import net.minecraft.core.component.DataComponentMap;
-import net.minecraft.util.profiling.Profiler;
-import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraft.world.item.ItemDisplayContext;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -24,14 +15,12 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.List;
-
 //? if >=1.21.10
 import net.minecraft.client.renderer.SubmitNodeCollector;
 
-@MixinEnvironment(value = "client")
+@MixinEnvironment("client")
 @Mixin(ShieldSpecialRenderer.class)
-public abstract class ShieldSpecialRendererMixin {
+abstract class ShieldSpecialRendererMixin {
 	@Shadow
 	@Final
 	private ShieldModel model;
@@ -45,31 +34,7 @@ public abstract class ShieldSpecialRendererMixin {
 			)
 	)
 	private void submitTrim(DataComponentMap argument, ItemDisplayContext displayContext, PoseStack poseStack, SubmitNodeCollector nodeCollector, int packedLight, int packedOverlay, boolean hasFoil, int outlineColor, CallbackInfo ci) {
-		ProfilerFiller profiler = Profiler.get();
-		profiler.push("trimica:shield");
-		List<DynamicTrimTextureAtlasSprite> dynamicSprites = TrimicaClient.getRuntimeAtlases().getShieldSprites(Minecraft.getInstance().level, argument);
-		for (DynamicTrimTextureAtlasSprite dynamicSprite : dynamicSprites) {
-			TrimPalette palette = dynamicSprite.getPalette();
-			int light = palette == null ? packedLight : (palette.isEmissive() ? LightTexture.FULL_BRIGHT : packedLight);
-			if (palette != null && palette.isAnimated()) {
-				Compat.ifSodiumPresent(compat -> compat.markSpriteAsActive(dynamicSprite));
-			}
-			nodeCollector.submitModelPart(
-					this.model.plate(),
-					poseStack,
-					dynamicSprite.getRenderType(),
-					light,
-					packedOverlay,
-					dynamicSprite,
-					false,
-					hasFoil,
-					-1,
-					null,
-					outlineColor
-			);
-		}
-
-		profiler.pop();
+		TrimicaClientApi.getInstance().getRenderer().submitShieldTrim(model.plate(), argument, poseStack, nodeCollector, packedLight, packedOverlay, hasFoil, outlineColor);
 	}
 	//?} else {
 	/*@Inject(
@@ -79,21 +44,8 @@ public abstract class ShieldSpecialRendererMixin {
 					target = "Lcom/mojang/blaze3d/vertex/PoseStack;popPose()V"
 			)
 	)
-	private void renderTrim(DataComponentMap dataComponentMap, ItemDisplayContext itemDisplayContext, PoseStack poseStack, MultiBufferSource multiBufferSource, int i, int j, boolean bl, CallbackInfo ci) {
-		ProfilerFiller profiler = Profiler.get();
-		profiler.push("trimica:shield");
-		List<DynamicTrimTextureAtlasSprite> dynamicSprites = TrimicaClient.getRuntimeAtlases().getShieldSprites(Minecraft.getInstance().level, dataComponentMap);
-		for (DynamicTrimTextureAtlasSprite dynamicSprite : dynamicSprites) {
-			TrimPalette palette = dynamicSprite.getPalette();
-			int light = palette == null ? i : (palette.isEmissive() ? LightTexture.FULL_BRIGHT : i);
-			if (palette != null && palette.isAnimated()) {
-				Compat.ifSodiumPresent(compat -> compat.markSpriteAsActive(dynamicSprite));
-			}
-			VertexConsumer vertexConsumer = dynamicSprite.wrap(ItemRenderer.getFoilBuffer(multiBufferSource, dynamicSprite.getRenderType(), itemDisplayContext == ItemDisplayContext.GUI, bl));
-			this.model.plate().render(poseStack, vertexConsumer, light, j);
-		}
-
-		profiler.pop();
+	private void renderTrim(DataComponentMap dataComponentMap, ItemDisplayContext itemDisplayContext, PoseStack poseStack, MultiBufferSource multiBufferSource, int packedLight, int packedOverlay, boolean hasFoild, CallbackInfo ci) {
+		TrimicaClientApi.getInstance().getRenderer().renderShieldTrim(model.plate(), dataComponentMap, itemDisplayContext, poseStack, multiBufferSource, packedLight, packedOverlay, hasFoild);
 	}
 	*///?}
 }